# docs/workflow/workflow-page-api-spec.yaml
openapi: 3.0.0
info:
  title: 工作流详情页 API
  version: 1.0.0
  description: |
    工作流详情页（/workflow/:id）相关的所有后端 API 接口定义。

    ## 页面功能概述

    工作流详情页是工作流编辑器的核心页面，支持以下功能：
    - 工作流基本信息管理（名称、描述、触发配置）
    - 工作流画布编辑（节点、连线、关联）
    - 工作流执行与状态监控
    - 执行历史查看
    - 工作流发布/取消发布

    ## API 分类

    1. **WorkflowController** - 工作流基础管理
    2. **WorkflowExecutionController** - 工作流执行管理
    3. **WorkflowNodeController** - 节点管理
    4. **WorkflowConnectionController** - 连线管理
    5. **WorkflowValidationController** - 验证与辅助接口
servers:
  - url: http://localhost:8080
    description: 后端服务地址
paths:
...
  /api/workflows/{workflowId}/validate:
    post:
      summary: 验证工作流
      description: |
        验证工作流的结构和配置是否正确，包括：
        - 节点连接是否完整（无孤立节点）
        - 节点参数配置是否正确
        - 变量引用是否有效
        - 循环引用检测

        ## 使用场景
        - 执行工作流前进行验证
        - 保存工作流时进行验证
      operationId: validateWorkflow
      tags:
        - WorkflowValidationController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 验证结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
  /api/workflows/{workflowId}/predecessors/{nodeUuid}:
    get:
      summary: 获取前置节点
      description: |
        获取指定节点的所有前置节点（上游节点）。

        ## 使用场景
        - 在节点配置面板中显示可引用的上游变量
      operationId: getPredecessorNodes
      tags:
        - WorkflowValidationController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeUuid
          in: path
          required: true
          description: 节点UUID
          schema:
            type: string
      responses:
        '200':
          description: 前置节点列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeDTO'
  /api/workflows/{workflowId}/available-variables/{nodeUuid}:
    get:
      summary: 获取节点可用变量
      description: |
        获取指定节点可以引用的所有变量列表。

        ## 使用场景
        - 在变量选择器中显示可用的变量
        - 配置节点参数时提供变量提示
      operationId: getAvailableVariables
      tags:
        - WorkflowValidationController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeUuid
          in: path
          required: true
          description: 节点UUID
          schema:
            type: string
      responses:
        '200':
          description: 可用变量列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableVariable'
  /api/workflows/{workflowId}/check-reference/{nodeUuid}:
    get:
      summary: 检查参数引用有效性
      description: |
        检查指定节点中的参数引用是否有效。

        ## 使用场景
        - 实时验证节点配置中的变量引用
      operationId: checkReference
      tags:
        - WorkflowValidationController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: nodeUuid
          in: path
          required: true
          description: 节点UUID
          schema:
            type: string
      responses:
        '200':
          description: 引用检查结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceCheckResult'
  /api/workflows/{workflowId}/execution-order:
    get:
      summary: 获取节点执行顺序
      description: |
        计算并返回工作流中节点的执行顺序。

        ## 使用场景
        - 在画布上显示执行顺序编号
        - 调试工作流执行流程
      operationId: getExecutionOrder
      tags:
        - WorkflowValidationController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 执行顺序
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: 节点UUID（按执行顺序排列）
components:
  schemas:
...
    NodeDTO:
      type: object
      description: |
        节点信息

        节点是工作流的基本组成单元，代表一个处理步骤。每个节点包含：
        - 唯一标识（nodeUuid）
        - 节点类型（type）
        - 位置信息（positionX, positionY）
        - 端口配置（inputPorts, outputPorts）
        - 参数定义（inputParams, outputParams）
        - 运行时配置（config）

        **注意**：`inputPorts`、`outputPorts`、`inputParams`、`outputParams`、`config` 字段都是 JSON 字符串格式。
      properties:
        id:
          type: integer
          format: int64
          description: 节点数据库ID（后端生成，保存时无需传入）
        nodeUuid:
          type: string
          description: |
            节点UUID，由前端生成的唯一标识符。
            格式示例：`node-start`、`node-loop`、`node-text-clean`
          example: "node-start"
        type:
          type: string
          description: |
            节点类型编码，常见类型：
            - `start`: 开始节点，定义工作流输入参数
            - `end`: 结束节点，定义工作流输出
            - `loop`: 循环节点，遍历数组执行循环体
            - `loopBodyCanvas`: 循环体画布，包含循环内部节点
            - `textClean`: 文本清洗节点
            - `tableExtract`: 表格提取节点
            - `tableGenerate`: 表格生成节点
            - `judgeModel`: 裁判模型节点
            - `apiAuto`: API调用节点
          example: "start"
        name:
          type: string
          description: 节点名称，用于显示和变量引用（如 `${开始.output}`）
          example: "开始"
        positionX:
          type: integer
          description: 节点在画布上的X坐标
          example: 80
        positionY:
          type: integer
          description: 节点在画布上的Y坐标
          example: 303
        inputPorts:
          type: string
          description: |
            输入端口配置，JSON字符串格式。
            示例：`[{"id":"input-1","name":"输入"}]`
          example: '[{"id":"input-1","name":"输入"}]'
        outputPorts:
          type: string
          description: |
            输出端口配置，JSON字符串格式。
            示例：`[{"id":"output-1","name":"输出"}]`
          example: '[{"id":"output-1","name":"输出"}]'
        inputParams:
          type: string
          description: |
            输入参数定义，JSON字符串格式。
            示例：`[{"name":"file","type":"File","fileType":"Excel","required":true}]`
          example: '[]'
        outputParams:
          type: string
          description: |
            输出参数定义，JSON字符串格式。
            示例：`[{"name":"output","type":"Array","elementType":"Object"}]`
          example: '[{"name":"dict","type":"Dictionary","required":false,"dictionaryType":"Demo数据字典"}]'
        config:
          type: string
          description: |
            节点运行时配置，JSON字符串格式。
            包含节点执行所需的具体参数值，支持变量引用。
          example: '{}'
        parentNodeUuid:
          type: string
          nullable: true
          description: |
            父节点UUID，用于嵌套节点（如循环体内部的节点指向循环体画布）。
            顶级节点此值为 null。
          example: null

    # ==================== 连线相关 ====================
...
    ValidationResult:
      type: object
      description: 工作流验证结果
      properties:
        valid:
          type: boolean
          description: 是否验证通过
          example: true
        errors:
          type: array
          description: 错误列表
          items:
            type: object
            properties:
              type:
                type: string
                description: 错误类型
                example: "INVALID_REFERENCE"
              message:
                type: string
                description: 错误信息
                example: "节点 '文本清洗' 引用了不存在的变量 '${开始.invalid_param}'"
              nodeUuid:
                type: string
                description: 相关节点UUID
                example: "node-text-clean"
    AvailableVariable:
      type: object
      description: 可用变量信息
      properties:
        name:
          type: string
          description: 变量名称
          example: "开始.dict"
        type:
          type: string
          description: 变量类型
          example: "Dictionary"
        sourceNodeName:
          type: string
          description: 来源节点名称
          example: "开始"
        description:
          type: string
          description: 变量描述
          example: "Demo数据字典"
    ReferenceCheckResult:
      type: object
      description: 引用检查结果
      properties:
        valid:
          type: boolean
          description: 引用是否有效
          example: true
        invalidReferences:
          type: array
          description: 无效的引用列表
          items:
            type: string
          example: []